Refactor code into functions for better modularity
Refactored the code for retrieving road network data and features for a selected city into separate function. This was done to improve the modularity and reusability of the code, allowing it to be used on any city of the world. The functions were tested on Kampala and Mexico City to esure that they produce the same results as the original code. This update makes the code more maintainable and scalable for future development.