Add support for panoramic images and higher resolution downloads
Updated the code for downloading images to support both panoramic and non-panoramic images. Non-panoramic images are downloaded with a resolution of 1024px, while panoramic images are downloaded with a resolution of 2048px and then cut into 4 separate images. The code need to be further improved to ensure that the resulting crops have the same size as non-panoramic images.