Change paths to save results to new directory
I updated the paths in the code to save the results of the roads,  selected points and seleted features to a new directory. This was done to better organize the project and make it easier to access and analyze the results. No changes were made to the functionality of the code, only the file paths were modified.