Add code to find the center of the roads in street images
Updated the code to include a new feature that finds the center of the roads in a given street image. The updated code first saves the segmentations of the street images as pickle files. It then uses these segmentation results to locate the center of the road or roads (if the image is panoramic). If the image has roads, it is considered as usable for analysis.

Some functions were taken from Matthew Danish's GitHub repository to help with the implementation of this feature. These functions were adapted to fit the specific needs of this project.