Improve workflow to incorporate panoramic image cropping in segmentation results
Enhanced the code workflow to enable cropping of panoramic images after segmentation and reflect the cropping in the segmentation results. This improvement allows us to use the final version of the images for calculating the GVI (Green View Index).  The cropping algorithm is currently very simple and not working for all the cases, so it needs to be improved.