Create Python scripts for image processing and parallel dataset splitting
Created two Python scripts to streamline the image processing workflow and enable parallel processing of the dataset. The first script contains all the necessary code to download the images, perform image segmentation, identify usable images, and calculate the GVI (Green View Index). This script serves as the main processing script, handling the entire image processing pipeline.

The second script focuses on splitting the dataset into smaller subsets, allowing parallel processing of the images. By dividing the dataset into manageable portions, the script maximizes computational efficiency by running multiple instances of the main processing script simultaneously. At the end of the parallel processing, the results are combined and stored in a GeoPackage (gpkg) file, providing a consolidated output for further analysis.

To execute the main script, use the following command in the terminal:
python main_script.py "city" "access_token" "path_to_points_file" "path_to_folder"

Note that the "path_to_folder" parameter is used to store the images in a directory. If it is left empty (""), the images are not going to be locally stored.

A successful test run was performed with 1,000 points in Kampala, Uganda. The script was executed with 8 processes, each with 5 workers, resulting in a total processing time of 1 hour and 10 minutes.