Fix bug in crop_panoramic_images function
Addressed a bug in the `crop_panoramic_images` function where the calculation of the right coordinate was incorrect. Previously, the width of the image was mistakenly used instead of the center coordinate, resulting in an error message stating that the "right" coordinate was less than the "left" coordinate.

The bug has been resolved by correctly calculating the right coordinate based on the center coordinate, ensuring that the cropping process for panoramic images functions as intended.