Parallel image processing only using with threading and result logging
This commit introduces the use of only threading to enable parallel processing of images (the multiprocessing library is not longer used). Additionally, a mechanism has been implemented to store the results in a CSV file after each completed process. This ensures that in the event of a process interruption or script termination, the calculations performed up until that point are preserved and readily available for future analysis. Furthermore, to ensure data integrity and prevent conflicts when accessing the CSV file for result storage, a lock mechanism has been implemented.