Updated README with environment setup instructions and Green View Index steps
Added detailed instructions to the README file explaining how to set up the environment to run the project in both a local environment and Google Colab. Included installation steps for required libraries and mounting Google Drive in Google Colab. Also, provided information about the necessary steps to obtain the Green View Index in the place specified by the user. These updates aim to make it easier for users to get started with the project.