Created Jupyter notebook for Google Colab integration
This commit adds a Jupyter notebook that simplifies the usage of the project by allowing users to run each cell of the notebook without the need to manually download files from the GitHub repository and upload them to Google Colab. The instructions to use the notebook are provided in the Readme file.